First pass refactoring console output
